create density and scattering submodules (#24). reorganization in preparation for PR (#29). Density representations no longer have a scatter routine.
