+++ src/cryojax/io/load_atoms.py
@@ -1,0 +1,5 @@
+"""
+Routines for loading atomic structures.
+"""
+
+__all__ = []
