+++ src/cryojax/simulator/scattering/__init__.py
@@ -1,0 +1,13 @@
+from . import base, gaussian_mixture, fourier_slice, nufft
+
+from .base import *
+from .fourier_slice import *
+from .nufft import *
+from .gaussian_mixture import *
+
+__all__ = (
+    base.__all__
+    + fourier_slice.__all__
+    + nufft.__all__
+    + gaussian_mixture.__all__
+)
